Perfil de usuario hecho y más funciones
se ha implementado tanto la página de perfil (a mitad) como el cambio de rol de los usuarios como la eliminación del perfil de usuario a través de la página "userProfile"

+++ Backend/Splitzy/Splitzy/Services/UserService.cs
@@ -15,7 +15,7 @@ public class UserService
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<User> GetByIdAsync(long id)
+    public async Task<User> GetByIdAsync(int id)
     {
         User user = await _unitOfWork.UserRepository.GetByIdAsync(id);
         return user;
